refactor: extract shared loadSource and detail metadata helpers
The three detail routes (components, blocks, templates) each carried a
byte-identical loadSource function and a near-identical generateMetadata
body. Extract loadSource to lib/registry-source.ts and the shared OG/Twitter
metadata shape to a new detailMetadata() helper in lib/site.ts, parameterized
by an optional titleSuffix. Verified the refactor is behavior-preserving: all
143 detail-page HTML outputs (og/twitter meta, canonical, source blocks) are
byte-identical before/after, modulo non-deterministic build-hash script tags.
