+++ app/(showcase)/components/[category]/[component]/page.tsx
@@ -10,12 +10,12 @@ import type {
   SourceFile,
 } from "@/components/showcase/component-page";
 import { highlightCode, highlightInline, langFromPath } from "@/lib/highlight";
-import { SITE } from "@/lib/site";
+import { loadSource } from "@/lib/registry-source";
+import { detailMetadata } from "@/lib/site";
 import {
   CATEGORY_LABELS,
   COMPONENTS,
   REGISTRY_BY_NAME,
-  entryHref,
   getExamples,
 } from "@/registry/hirael/registry-meta";
 import registryProps from "@/registry/hirael/registry-props.json";
@@ -43,64 +43,7 @@ export async function generateMetadata({
     entry.category !== category
   )
     return {};
-  const href = entryHref(entry);
-  const url = `${SITE.url}${href}`;
-  const title = `${entry.title} - ${SITE.name}`;
-  return {
-    title: entry.title,
-    description: entry.description,
-    alternates: {
-      canonical: href,
-    },
-    openGraph: {
-      type: "article",
-      url,
-      siteName: SITE.name,
-      title,
-      description: entry.description,
-      images: [
-        {
-          url: "/opengraph-image",
-          width: 1200,
-          height: 630,
-          alt: title,
-        },
-      ],
-    },
-    twitter: {
-      card: "summary_large_image",
-      title,
-      description: entry.description,
-      images: ["/opengraph-image"],
-    },
-  };
-}
-
-async function loadSource(
-  files: string[] | undefined,
-): Promise<Record<string, SourceFile>> {
-  const out: Record<string, SourceFile> = {};
-  if (!files) return out;
-  await Promise.all(
-    files.map(async (f) => {
-      const abs = path.join(process.cwd(), f);
-      let code: string;
-      try {
-        code = await fs.readFile(abs, "utf8");
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : String(err);
-        // Surface the real reason — silently rendering the stub made it
-        // very hard to tell whether the path was wrong, the file was
-        // missing, or the deploy bundle didn't include it.
-        console.error(`[loadSource] could not read ${abs}: ${msg}`);
-        code = `// (unable to read source: ${msg})`;
-      }
-      const lang = langFromPath(f);
-      const html = await highlightCode(code, lang);
-      out[f] = { code, html, lang };
-    }),
-  );
-  return out;
+  return detailMetadata(entry);
 }
 
 /**
