+++ lib/site.ts
@@ -2,6 +2,11 @@
  * Site-wide brand constants. One source of truth for name, URLs, social links.
  */
 
+import type { Metadata } from "next";
+
+import type { RegistryEntryMeta } from "@/registry/hirael/registry-meta";
+import { entryHref } from "@/registry/hirael/registry-meta";
+
 export const SITE = {
   name: "Hirael",
   fullName: "Hirael",
@@ -44,3 +49,52 @@ export const NAV_LINKS: { href: string; label: string; external?: boolean }[] =
     { href: "/templates", label: "Templates" },
     { href: "/theme", label: "Theme" },
   ];
+
+/**
+ * Shared `generateMetadata` body for the component/block/template detail
+ * routes — same title, OpenGraph article, and Twitter card shape, differing
+ * only in the title suffix. Reproduces each route's pre-existing title
+ * format exactly (bare title dash-joined with the site name for components;
+ * "<suffix> | Hirael" for blocks/templates) so this extraction changes no
+ * rendered output.
+ */
+export function detailMetadata(
+  entry: RegistryEntryMeta,
+  opts: { titleSuffix?: string } = {},
+): Metadata {
+  const { titleSuffix } = opts;
+  const href = entryHref(entry);
+  const url = `${SITE.url}${href}`;
+  const pageTitle = titleSuffix ? `${entry.title} ${titleSuffix}` : entry.title;
+  const ogTitle = titleSuffix
+    ? `${pageTitle} | ${SITE.name}`
+    : `${pageTitle} - ${SITE.name}`;
+  return {
+    title: pageTitle,
+    description: entry.description,
+    alternates: {
+      canonical: href,
+    },
+    openGraph: {
+      type: "article",
+      url,
+      siteName: SITE.name,
+      title: ogTitle,
+      description: entry.description,
+      images: [
+        {
+          url: "/opengraph-image",
+          width: 1200,
+          height: 630,
+          alt: ogTitle,
+        },
+      ],
+    },
+    twitter: {
+      card: "summary_large_image",
+      title: ogTitle,
+      description: entry.description,
+      images: ["/opengraph-image"],
+    },
+  };
+}
